+++ CHANGELOG.md
@@ -8,6 +8,12 @@ This change log follows the format documented in [Keep a CHANGELOG].
 [semantic versioning]: http://semver.org/
 [keep a changelog]: http://keepachangelog.com/
 
+## v0.2.1 - 2024-05-27
+
+### Fixed
+
+- Fixed `FromCoercer` type to properly resolve the type shape.
+
 ## v0.2.0 - 2024-05-27
 
 ### Fixed